second fixes

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: wkhosa <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/17 10:53:52 by wkhosa            #+#    #+#             */
-/*   Updated: 2017/08/15 14:52:13 by wkhosa           ###   ########.fr       */
+/*   Updated: 2017/09/05 16:12:20 by wkhosa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int		ft_atoi(const char *str)
 {
-	unsigned long	int	nbr;
-	int					index;
-	int					sign;
+	int		nbr;
+	int		index;
+	int		sign;
 
 	nbr = 0;
 	sign = 1;
@@ -32,6 +32,8 @@ int		ft_atoi(const char *str)
 	while (str[index] >= '0' && str[index] <= '9')
 	{
 		nbr = (str[index] - '0')  + nbr * 10;
+	//	if (nbr >= 2147483648)
+	//		return (-2147483648);
 		index++;
 	}
 /*
@@ -40,5 +42,5 @@ int		ft_atoi(const char *str)
 	if (nbr <= -214743648)
 		nbr = -214743648;
 */
-	return ((signed int)nbr * sign);
+	return ((int)nbr * sign);
 }
